storageTable 이동평균법 추가(1차)

+++ dashboard-spring/src/main/java/com/humintecTest/dashboard/vo/StorageTableVo.java
@@ -8,6 +8,6 @@ import java.sql.Date;
 @Data
 public class StorageTableVo {
     private Date date;
-    private BigInteger used;
+    private Double used;
     private String storage_name;
 }
